📦️ [chore] #518 : common 디렉토리 생성 및 컴포넌트 분리
