+++ client/src/App.tsx
@@ -6,8 +6,8 @@ import { ErrorBoundary } from "react-error-boundary";
 import GlobalStyle from "@styles/GlobalStyle";
 import { RoutePath } from "@constants/enums";
 import { DEFAULT_STALE_TIME } from "@constants/consts";
-import Loading from "@components/Loading";
-import ErrorFallback from "@components/ErrorFallback";
+import Loading from "src/common/layer/Loading";
+import ErrorFallback from "src/common/layer/ErrorFallback";
 import NotFoundPage from "@pages/NotFountPage";
 import Exception from "./services/Exception";
 import { authStorage } from "./services/ClientStorage";
