+++ client/src/common/design/CardsScroller/style.ts
