+++ client/src/common/design/Modal/style.ts
