--- client/src/components/BottomNavigationButton/style.ts
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-import { NavLink } from "react-router-dom";
-
-export const LinkButton = styled(NavLink)`
-  width: 35px;
-  filter: ${({ theme }) => theme.COLORS.FILTER_GRAY};
-  &.active {
-    filter: ${({ theme }) => theme.COLORS.FILTER_BLUE};
-  }
-  & > img {
-    width: 100%;
-  }
-`;
