+++ client/src/components/Calendar/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState, Dispatch, SetStateAction } from "react";
 import { useNavigate } from "react-router-dom";
 import dayjs from "dayjs";
 import weekOfYear from "dayjs/plugin/weekOfYear";
-import Paper from "@components/design/Paper";
+import Paper from "src/common/design/Paper";
 import CalendarHeader from "@components/Calendar/CalendarHeader";
 import CalendarBody from "@components/Calendar/CalendarBody";
 import { RoutePath } from "@constants/enums";
