+++ client/src/components/CalendarHeatMap/index.tsx
@@ -8,7 +8,7 @@ import {
 import { FormatMonth, ExerciseState } from "@constants/enums";
 import getCalendarHeatMapArray from "@utils/getCalendarHeatMapArray";
 import getExerciseStateForOneYear from "@utils/getExerciseStateForOneYear";
-import Paper from "@components/design/Paper";
+import Paper from "src/common/design/Paper";
 import useAllExerciseDate from "@hooks/query/exercise/useAllExerciseDate";
 import { authStorage } from "src/services/ClientStorage";
 import { UserId } from "src/types/user";
