+++ client/src/pages/PageTemplate.tsx
@@ -1,9 +1,9 @@
 import React, { Suspense } from "react";
 import { ToastContainer } from "react-toastify";
-import TopNavigationBar from "@components/TopNavigationBar";
-import MainContainer from "@components/MainContainer";
-import BottomNavigationBar from "@components/BottomNavigationBar";
-import Loading from "@components/Loading";
+import TopNavigationBar from "src/common/layer/TopNavigationBar";
+import MainContainer from "src/common/layer/MainContainer";
+import BottomNavigationBar from "src/common/layer/BottomNavigationBar";
+import Loading from "src/common/layer/Loading";
 import checkIsIOS from "@utils/checkIsIOS";
 import checkIsPWADisplayMode from "@utils/checkIsPWADisplayMode";
 import "react-toastify/dist/ReactToastify.css";
